DeskClock: Don't show duplicate dialog when resume to activity
It is the NullPointerException which cause the application died.
To call the method to show delete confirm dialog when resume to
activity is not the right time.

The method to show delete confirm dialog will be called by
initialize method which means we will call it only when activity
is recreated.

CRs-Fixed: 523897

Change-Id: I38639c163e50b437d39bab62667e33d827dab35c