DeskClock: Alarm clock disturbs incoming call ring
This feature is required by CMCC. When a incoming call is ringing,
the alarm wakes up on time. The alarm clock shall not disturb the
call ring with any sound or UI. The alarm clock shall snooze until
the call is ended. As soon as the call is hanged, the alarm view
shall resume it self.

Remove the LOCAL_SDK_VERSION from makefile to use SystemProperties.
"persist.env.deskclock.incallsnz" is used as a switch for this feature.
While the phone is busy for the calls, shortcut the alarm intent and
don't make the alarm snooze until the phone call status is idle.

CRs-Fixed: 543890

Change-Id: I685c5d7f5ceb6bbb30eea85abe376d7944d374dd