DeskClock: Fix alarm repeat infomation display not right when new an alarm
When user collapse the alarm, it will show the label and repeat infomation
on bootom, but this area is not update layout before draw, this will lead
the view use previous width and hight, So the infomation maybe not
full display.

When user collapse alarm, request the repeat infomations layout.

CRs-fixed: 620426

Change-Id: If6dcf03d22286b2dc8fcfeb267f06a8229362ddf