DeskClock: The alarm ringtone name display error
Select the downloaded music,there is no operation to get the name
from the uri.If restart Alarm,there is no permission to get the
title from the uri.

No matter in which database,the music has the same id.We can only
get the info of the music from other database by id in uri.
Then set the name to the alarm.

CRs-fixed: 662043

Change-Id: Id284eb9b6156d5c0c45e228c56ccc3c707d0c7c2