DeskClock: Fix the reset icon can't worked when setting value as 9h99m9s
When user set timer to 9h99m99s, the clocktick will  set the +1 button
disable, because the +1 button is same with reset button, when user pause
the timer, the reset button also be disable, so it will no response.

Set the reset button to enable when pause the timer.

Change-Id: Ib6d2546039720ea8bd94f8c139724636e012187d
CRs-fixed: 865250